server: add echo server using io_uring
The echo server is written using direct syscall to interact with io_uring
without liburing (for education purpose). This is the first working version
and we still need to profile to optimize more as the benchmark result is
not good.

The benchmark result on localhost interface:
./client --client 32 --packet.size 32000 --duration 30s
Sent: 65973.76 req/s. Received: 65972.70 req/s
DataSent: 1.97 GB/s. Received: 1.97 GB/s